requirements.txt
# 基础依赖
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# 飞书SDK
lark-oapi>=1.0.0

# 情感分析
snownlp>=0.12.3

# 配置管理
pyyaml>=6.0

# 定时任务
schedule>=1.2.0

# 浏览器自动化（小红书/微信登录）
playwright>=1.40.0

# Web框架
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
